Reorganized the code into separate passmanager and passes
